% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_functions.R
\name{soma}
\alias{soma}
\title{Soma dois números.}
\usage{
soma(x, y)
}
\arguments{
\item{x}{Um número.}

\item{y}{Outro número.}
}
\value{
A soma de x e y.
}
\description{
Soma dois números.
}
\examples{
soma(2, 3)

}
